refactor: move controllers and services to their own foldres
